requirements.txt
aiogram>=3.4.1
telethon>=1.36.0
python-dotenv>=1.0.1
playwright>=1.40.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.4
